feat: utils 기반 코드-> fast api 기반 project structure (schema, model, crud, services, api) 마이그레이션
