+++ services/monitoring/__init__.py
@@ -1,0 +1,3 @@
+from services.monitoring.service import MonitoringService
+
+__all__ = ["MonitoringService"]
